requirements.txt
# Core dependencies
google-generativeai==0.8.5
pymongo==4.13.2
python-dotenv==1.0.0

# Optional: Jupyter support (if you want to keep the notebook)
jupyter==1.0.0
ipykernel==6.25.0

# Development dependencies (optional)
pytest==7.4.0
black==23.7.0
flake8==6.0.0